move atomic exceptions
