+++ plasmapy/atomic/atomic.py
@@ -17,12 +17,7 @@ from plasmapy.atomic.particle_class import Particle
 from plasmapy.atomic.particle_input import particle_input
 from plasmapy.atomic.symbols import atomic_symbol
 
-from plasmapy.utils import (
-    InvalidParticleError,
-    InvalidElementError,
-    InvalidIsotopeError,
-    MissingAtomicDataError,
-)
+from plasmapy.atomic.exceptions import MissingAtomicDataError, InvalidIsotopeError, InvalidElementError, InvalidParticleError
 
 from numbers import (
     Integral,
