+++ plasmapy/atomic/ionization_states.py
@@ -11,7 +11,8 @@ import numpy as np
 import astropy.units as u
 
 from plasmapy.atomic import atomic_number, Particle, particle_symbol, IonizationState, State
-from plasmapy.utils import AtomicError, ChargeError, InvalidParticleError, check_quantity
+from plasmapy.utils import check_quantity
+from plasmapy.atomic.exceptions import AtomicError, ChargeError, InvalidParticleError
 
 __all__ = ["IonizationStates"]
