+++ plasmapy/atomic/tests/test_atomic.py
@@ -38,15 +38,9 @@ from ..nuclear import (
 
 from ...utils import (
     run_test,
-    AtomicWarning,
-    InvalidElementError,
-    InvalidIsotopeError,
-    InvalidIonError,
-    ChargeError,
-    AtomicError,
-    MissingAtomicDataError,
-    InvalidParticleError,
-)
+    )
+from plasmapy.atomic.exceptions import AtomicError, MissingAtomicDataError, ChargeError, InvalidIonError, InvalidIsotopeError, InvalidElementError, \
+    InvalidParticleError, AtomicWarning
 
 # function to be tested, argument(s), expected result/outcome
