+++ plasmapy/atomic/tests/test_ionization_states.py
@@ -7,7 +7,8 @@ import itertools
 import numpy as np
 import astropy.units as u
 
-from plasmapy.utils import AtomicError, InvalidIsotopeError, run_test
+from plasmapy.utils import run_test
+from plasmapy.atomic.exceptions import AtomicError, InvalidIsotopeError
 from plasmapy.atomic import (
     State,
     IonizationState,
