+++ plasmapy/atomic/tests/test_nuclear.py
@@ -2,12 +2,10 @@ from astropy import units as u, constants as const
 import numpy as np
 from ..nuclear import nuclear_binding_energy, nuclear_reaction_energy, mass_energy
 from ...utils import (
-    InvalidParticleError,
-    InvalidIsotopeError,
-    AtomicError,
     run_test,
     run_test_equivalent_calls,
 )
+from plasmapy.atomic.exceptions import AtomicError, InvalidIsotopeError, InvalidParticleError
 import pytest
 
 test_nuclear_table = [
