+++ plasmapy/physics/parameters.py
@@ -34,7 +34,8 @@ import numpy as np
 from plasmapy.constants import (m_p, m_e, c, mu0, k_B, e, eps0, pi)
 from plasmapy import atomic, utils
 
-from plasmapy.utils.exceptions import (PhysicsError, AtomicError)
+from plasmapy.utils.exceptions import (PhysicsError)
+from plasmapy.atomic.exceptions import AtomicError
 
 
 def _grab_charge(ion, z_mean=None):
