+++ plasmapy/transport/tests/test_transport.py
@@ -7,8 +7,8 @@ from plasmapy.atomic.atomic import particle_mass, integer_charge
 from astropy.tests.helper import assert_quantity_allclose
 from plasmapy.utils.exceptions import (PhysicsError,
                                        PhysicsWarning,
-                                       InvalidParticleError,
                                        RelativityWarning)
+from plasmapy.atomic.exceptions import InvalidParticleError
 from plasmapy.physics.parameters import Hall_parameter
 from plasmapy.constants import m_p, m_e
