+++ plasmapy/utils/__init__.py
@@ -6,20 +6,13 @@ from .checks import (check_quantity,
 from .exceptions import (PlasmaPyError,
                          PhysicsError,
                          RelativityError,
-                         AtomicError,
-                         MissingAtomicDataError,
-                         ChargeError,
-                         InvalidIonError,
-                         InvalidIsotopeError,
-                         InvalidElementError,
-                         InvalidParticleError,
                          DataStandardError,
                          PlasmaPyWarning,
                          PhysicsWarning,
                          CouplingWarning,
-                         RelativityWarning,
-                         AtomicWarning,
-                         MissingAtomicDataWarning)
+                         RelativityWarning)
+from plasmapy.atomic.exceptions import AtomicError, MissingAtomicDataError, ChargeError, InvalidIonError, InvalidIsotopeError, InvalidElementError, \
+    InvalidParticleError, AtomicWarning, MissingAtomicDataWarning
 
 from .pytest_helpers import (
     run_test,
